# core: don't allocate in DedupAddrs