--- a/core/network/network_test.go
+++ b/core/network/network_test.go
@@ -2,6 +2,8 @@ package network
 
 import (
 	"fmt"
+	"math"
+	"math/rand"
 	"testing"
 
 	ma "github.com/multiformats/go-multiaddr"
@@ -30,6 +32,27 @@ func TestDedupAddrs(t *testing.T) {
 			deduped := DedupAddrs(tc.in)
 			for _, a := range tc.out {
 				require.Contains(t, deduped, a)
+			}
+		})
+	}
+}
+
+func BenchmarkDedupAddrs(b *testing.B) {
+	b.ReportAllocs()
+	var addrs []ma.Multiaddr
+	r := rand.New(rand.NewSource(1234))
+	for i := 0; i < 100; i++ {
+		tcpAddr := ma.StringCast(fmt.Sprintf("/ip4/127.0.0.1/tcp/%d", r.Intn(math.MaxUint16)))
+		quicAddr := ma.StringCast(fmt.Sprintf("/ip4/127.0.0.1/udp/%d/quic-v1", r.Intn(math.MaxUint16)))
+		wsAddr := ma.StringCast(fmt.Sprintf("/ip4/127.0.0.1/tcp/%d/ws", r.Intn(math.MaxUint16)))
+		addrs = append(addrs, tcpAddr, tcpAddr, quicAddr, quicAddr, wsAddr)
+	}
+	for _, sz := range []int{10, 20, 30, 50, 100} {
+		b.Run(fmt.Sprintf("%d", sz), func(b *testing.B) {
+			items := make([]ma.Multiaddr, sz)
+			for i := 0; i < b.N; i++ {
+				copy(items, addrs[:sz])
+				DedupAddrs(items)
 			}
 		})
 	}