Closes #2352.