@marten-seemann 2023-06-27
Benchmarking result (compared to what we have in master):
```
name               old time/op    new time/op    delta
DedupAddrs/10-10      396ns ± 1%     178ns ± 2%   -54.93%  (p=0.016 n=4+5)
DedupAddrs/20-10     1.33µs ± 2%    0.62µs ± 0%   -53.35%  (p=0.016 n=4+5)
DedupAddrs/30-10     1.67µs ± 1%    0.92µs ± 1%   -44.70%  (p=0.008 n=5+5)
DedupAddrs/50-10     3.21µs ± 5%    1.85µs ± 3%   -42.30%  (p=0.008 n=5+5)
DedupAddrs/100-10    8.61µs ± 6%    4.92µs ±13%   -42.84%  (p=0.008 n=5+5)

name               old alloc/op   new alloc/op   delta
DedupAddrs/10-10      88.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
DedupAddrs/20-10      88.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
DedupAddrs/30-10      88.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
DedupAddrs/50-10      88.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
DedupAddrs/100-10     88.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)

name               old allocs/op  new allocs/op  delta
DedupAddrs/10-10       3.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
DedupAddrs/20-10       3.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
DedupAddrs/30-10       3.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
DedupAddrs/50-10       3.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
DedupAddrs/100-10      3.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)